refactor: update CLI imports to new commands submodule structure
